Tests/LibWeb/Text/expected/SVG/svg-partial-relayout-does-not-leak-ancestor-paintables.txt

before=1
after=5
stable=false
